+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Holding alt while creating new sticky notes creates a hidden note.
+- There’s a macro with which you can create you own sticky note templates
+
 ## [1.2.1] - 2024-07-15
 ### Fixed
 - Fixed an issue where no sticky note could be placed.